+++ quality.sh
@@ -23,7 +23,7 @@ if [ -f "$tmp_path" ]; then
     rm "$tmp_path" || customDie "rm \"$tmp_path\" failed."
 fi
 echo > "$tmp_path"
-for name in example-cli.py setup.py testing.pyw pypicolcd/lcddaemon.py pypicolcd/command_line.py pypicolcd/__init__.py
+for name in example-cli.py setup.py testing.pyw pypicolcd/lcdframebuffer.py pypicolcd/command_line.py pypicolcd/__init__.py
 do
 echo "* checking '$name'..."
 pycodestyle-3 $name > "$tmp_path"